+++ R/lmranks_model_selection.R
@@ -5,8 +5,19 @@ logLik.lmranks <- function(object, ...){
 
 #' @export
 anova.lmranks <- function(object, ...){
-  cli::cli_abort("This method might not return correct results.")
-  NextMethod()
+  cli::cli_abort("This method is not implemented.")
+}
+
+#' @describeIn lmranks Projections of the data onto terms of rank-rank regression model
+#' @importFrom stats proj
+#' @inheritParams stats::proj
+#' @export
+proj.lmranks <- function(object, onedf = FALSE, ...){
+  object$df.residual <- stats::nobs(object) - length(coef(object))
+  out <- NextMethod()
+  object$df.residual <- NA
+  attr(out, "df") <- c(NA, NA, NA)
+  return(out)
 }
 
 #' @export