+++ R/ranks.R
@@ -2,7 +2,7 @@
 #'
 #' Marginal and simultaneous confidence sets for ranks.
 #'
-#' @param x vector of estimates containing estimated features by which the \code{length} populations are to be ranked.
+#' @param x vector of estimates containing estimated features by which the populations are to be ranked.
 #' @param Sigma estimated covariance matrix of \code{x}.
 #' @param coverage nominal coverage of the confidence set. Default is 0.95.
 #' @param cstype type of confidence set (\code{two-sided}, \code{upper}, \code{lower}). Default is \code{two-sided}.
@@ -31,7 +31,9 @@
 #' csranks(thetahat, Sigmahat)
 #' 
 #' # PISA example:
-#' attach(pisa)
+#' data(pisa2018)
+#' math_score <- pisa2018$math_score
+#' math_se <- pisa2018$math_se
 #' math_cov_mat <- diag(math_se^2)
 #' 
 #' # marginal confidence set for each country: